requirements.txt
pillow
openai
python-editor
